+++ packages/coding-standards/src/Sniffs/ForbiddenDumpSniff.php
@@ -1,0 +1,26 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Shopsys\CodingStandards\Sniffs;
+
+use PHP_CodeSniffer\Standards\Generic\Sniffs\PHP\ForbiddenFunctionsSniff;
+
+final class ForbiddenDumpSniff extends ForbiddenFunctionsSniff
+{
+    /**
+     * A list of forbidden functions with their alternatives.
+     *
+     * The value is NULL if no alternative exists. IE, the
+     * function should just not be used.
+     *
+     * @var string[]|null[]
+     */
+    public $forbiddenFunctions = [
+        'd' => null,
+        'dump' => null,
+        'print_r' => null,
+        'var_dump' => null,
+        'var_export' => null,
+    ];
+}